requirements.txt
python-jobspy
google-generativeai
pandas
